Classifications: Set all `CC-SA` to `copyleft-strong`
This commit sets all `CC-SA` style licenses to `copyleft-strong`.
With `property:no-modifications` and `property:non-commercial`,
any `ND` and `NC` properties are atteched to classified licenses.
So there is no need to keep `free-restricted` as the category of
the `CC-N*-SA` licenses. Instead the class `copyleft-strong` is
used to mark the copyleft obligation of a `SA` license.

Signed-off-by: Jens Keim <jens.keim@forvia.com>